requirements.txt
# Flask Backend
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# NLP
nltk==3.8.1

# Utilities
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.2
